+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 1.2.3
 * Monsters will now be able to despawn when farther than any player by 48 blocks (vanilla is 32) and will instantly despawn when farther away than 80 blocks (vanilla is 128) (configurable in "Base Feature")
 * Breaching creepers now breach more often
 * Launching creepers that go on cooldown now have higher cooldown each time they fail